+++ docs/demo/1.csrf/MyModule.php
@@ -7,7 +7,7 @@ class MyModule extends AbstractModule
 {
     protected function configure()
     {
-        $this->install(new AuraInputModule);
+        $this->install(new WebFormModule());
         $this->bind(FormInterface::class)->annotatedWith('contact_form')->to(ContactForm::class);
     }
 }